+++ test/presentation/pages/exercise_form/exercise_form_page_test.dart
@@ -3,12 +3,12 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:workout_log/domain/models/body_part.dart';
 import 'package:workout_log/domain/models/exercise.dart';
-import 'package:workout_log/presentation/pages/exercise_form_page.dart';
+import 'package:workout_log/presentation/pages/exercise_form/exercise_form_page.dart';
 import 'package:workout_log/presentation/providers/data_providers.dart';
 import 'package:workout_log/presentation/providers/selected_date_provider.dart';
 
-import '../helpers/test_app.dart';
-import '../test_helper.dart';
+import '../../../helpers/test_app.dart';
+import '../../../test_helper.dart';
 
 /// Sets the test surface to a portrait phone-like aspect (400 x 800)
 /// so the form's landscape branch — which contains a known 1px overflow
